streamlit_app/pyproject.toml
[project]
name = "pipool-dashboard"
version = "1.0.0"
description = "Streamlit web dashboard for PiPool automation system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "paho-mqtt>=2.0.0",
    "pandas>=2.0.0",
    "plotly>=5.18.0",
]

[dependency-groups]
dev = []
